--- packages/commonwealth/client/styles/pages/new_proposal_page.scss
@@ -1,133 +1,0 @@
-@import 'shared';
-
-.NewProposalPage {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  flex: 1;
-  padding: 24px;
-
-  h3 {
-    margin-top: 0;
-    margin-bottom: 20px;
-  }
-
-  @include extraSmall {
-    width: initial;
-  }
-
-  .topic-loading-spinner-wrap {
-    height: 120px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-}
-
-.NewProposalForm {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  max-width: 432px;
-  // @include mediumSmall {
-  //   max-width: 85%;
-  // }
-
-  textarea {
-    min-height: 200px;
-  }
-
-  .council-motion-description {
-    margin-top: -12px;
-    margin-bottom: 22px;
-    color: $neutral-500;
-  }
-
-  .cui-overlay-inline {
-    display: block;
-  }
-
-  .cui-grid {
-    max-width: 100%;
-
-    & > * {
-      max-width: 100%;
-    }
-  }
-}
-
-.AaveGovernance {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  .executor {
-    background-color: $neutral-50;
-    padding: 16px;
-    border-radius: $border-radius-corners;
-    border: 1px solid $neutral-100;
-    min-width: 100%;
-
-    .label {
-      font-size: 14px;
-      color: $neutral-600;
-    }
-
-    .mt-16 {
-      margin-top: 16px;
-    }
-  }
-
-  .selected-executor {
-    background-color: $green-50;
-    border: 1px solid $green-300;
-  }
-
-  .tabs {
-    margin-left: -15px;
-  }
-
-  .ml-12 {
-    margin-left: 12px;
-  }
-
-  .flex-label {
-    display: flex;
-    justify-content: space-between;
-    align-items: baseline;
-
-    .helper-text {
-      font-size: 14px;
-      color: $neutral-500;
-    }
-  }
-
-  .delegate-call-container {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-
-    .buttons-row {
-      display: flex;
-      gap: 16px;
-    }
-  }
-
-  .button {
-    border-radius: $border-radius-corners;
-    font-weight: 500;
-    color: $neutral-500;
-  }
-
-  .active {
-    border-color: $purple-300;
-    color: $purple-300;
-  }
-
-  .tab-selector {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-}
