+++ packages/commonwealth/client/styles/pages/new_snapshot_proposal.scss
@@ -1,0 +1,20 @@
+@import 'shared';
+
+.NewSnapshotProposalPage {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  gap: 16px;
+  flex: 1;
+  padding: 24px;
+  max-width: 456px;
+
+  @include extraSmall {
+    width: 100%;
+  }
+
+  .date-range {
+    display: flex;
+    flex-direction: column;
+  }
+}
